add routes, controller and service for user auth

+++ src/controllers/user/user.controller.ts
@@ -1,26 +1,18 @@
 import { NextFunction, Request, Response } from "express";
 import UserService from "../../services/user/user.service.js";
 import { User } from "../../common/types/user/user.type.js";
+import { JwtRequest } from "../../common/types/jwt/jwt-request.type.js";
+import HttpError from "../../helpers/error/http.error.js";
+import { HttpCode } from "../../common/enums/http/http-code.enum.js";
 
 class UserController {
     #userService: UserService;
 
     constructor(userService: UserService) {
         this.#userService = userService;
-        this.create = this.create.bind(this);
         this.getById = this.getById.bind(this);
         this.updateById = this.updateById.bind(this);
         this.deleteById = this.deleteById.bind(this);
-    }
-
-    public async create(req: Request, res: Response, next: NextFunction): Promise<void> {
-        try {
-            const user = await this.#userService.create(req.body as User);
-
-            res.send(user);
-        } catch(err) {
-            next(err);
-        }
     }
 
     public async getById(req: Request, res: Response, next: NextFunction): Promise<void> {
@@ -37,6 +29,14 @@ class UserController {
     public async updateById(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
             const { id } = req.params as Record<'id', string>;
+
+            if(id !== (req as JwtRequest).token.id) {
+                throw new HttpError({
+                    status: HttpCode.UNAUTHORIZED, 
+                    message: 'User id mismatch'
+                });
+            }
+
             const user = await this.#userService.updateById(id, req.body as User);
         
             res.send(user);
@@ -48,6 +48,14 @@ class UserController {
     public async deleteById(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
             const { id } = req.params as Record<'id', string>;
+
+            if(id !== (req as JwtRequest).token.id) {
+                throw new HttpError({
+                    status: HttpCode.UNAUTHORIZED, 
+                    message: 'User id mismatch'
+                });
+            }
+
             const user = await this.#userService.deleteById(id);
         
             res.send(user);
